Big Entities Update
+Added enemySpawn list on config for both levels and their correspondent enemies creation process.
+Reworked player and enemy workflow and factory methods slightly. Now controlled partially from scene.
+ Reordered Module Call order:
Others -> Scene -> fade- > render
+Separated debug keys on scene and player depending on functionality
+Added gamePause, transforms dt to 0 and stops entities and user interface logic updates.

+++ Get Hooked/Motor2D/j1Map.cpp
@@ -43,8 +43,8 @@ void j1Map::Draw()
 		Image = data.imagelayers.At(x)->data;
 
 		App->render->Blit(data.imagelayers[x]->texture,
-			Image->SpeedBack, data.imagelayers[x]->OffsetY,
-			&data.imagelayers[x]->GetImageLayerRect(), SDL_FLIP_NONE, 0U, data.imagelayers[x]->speed);
+			(int)Image->SpeedBack, (int)data.imagelayers[x]->OffsetY,
+			&data.imagelayers[x]->GetImageLayerRect(), SDL_FLIP_NONE, 0U, (uint)data.imagelayers[x]->speed);
 	}
 
 	p2List_item<TileSet*>* item; //Sprites_Layer
